Fix for token being cleared with FB login
After logging in with FB and a request to the Graph API failed,
componentDidMount would jump the the catch block and clear localStorage.
Solution was to put the request to the FB Graph API in it's own
try/catch block so a failed request wouldn't jump to the outer catch
block and clear localStorage.