Fetch and render friends on CountryPanel click
`updateCurrentCountry` method will call `getCountryFriends` if a `facebook`
prop exists on user object. It will then updated the getCurrentCountry
object on state with a `friends` prop that contains an array of friends
and the status for their country.
`FriendsList` will render any friends from on 'friends' property in
`currentCountry` if it exists.